feat(graph-sdk-types): implement base types for GraphSDK #4
