+++ core/src/dsnp/mod.rs
@@ -1,6 +1,6 @@
-mod compression;
+pub mod api_types;
+pub mod compression;
 pub mod dsnp_types;
-mod encryption;
-#[allow(dead_code)] // todo: remove after usage
-mod reader_writer;
-mod schema;
+pub mod encryption;
+pub mod reader_writer;
+pub mod schema;
