+++ core/src/frequency/mod.rs
@@ -1,0 +1,4 @@
+pub mod reader_writer;
+
+/// A utility to read/write data from and to Frequency chain specific implementation of DSNP
+pub struct Frequency;
