+++ core/src/lib.rs
@@ -1,2 +1,3 @@
 mod dsnp;
+mod frequency;
 mod types;
